+++ color/src/impl_bytemuck.rs
@@ -3,7 +3,10 @@
 
 #![allow(unsafe_code, reason = "unsafe is required for bytemuck unsafe impls")]
 
-use crate::{AlphaColor, ColorSpace, OpaqueColor, PremulColor, PremulRgba8, Rgba8};
+use crate::{
+    AlphaColor, ColorSpace, ColorSpaceTag, HueDirection, OpaqueColor, PremulColor, PremulRgba8,
+    Rgba8,
+};
 
 // Safety: The struct is `repr(transparent)` and the data member is bytemuck::Pod.
 unsafe impl<CS: ColorSpace> bytemuck::Pod for AlphaColor<CS> {}
@@ -44,10 +47,62 @@ unsafe impl bytemuck::Pod for Rgba8 {}
 // Safety: The struct is `repr(C)` and all members are bytemuck::Zeroable.
 unsafe impl bytemuck::Zeroable for Rgba8 {}
 
+// Safety: The enum is `repr(u8)` and has only fieldless variants.
+unsafe impl bytemuck::NoUninit for ColorSpaceTag {}
+
+// Safety: The enum is `repr(u8)` and `0` is a valid value.
+unsafe impl bytemuck::Zeroable for ColorSpaceTag {}
+
+// Safety: The enum is `repr(u8)`.
+unsafe impl bytemuck::checked::CheckedBitPattern for ColorSpaceTag {
+    type Bits = u8;
+
+    fn is_valid_bit_pattern(bits: &u8) -> bool {
+        use bytemuck::Contiguous;
+        // Don't need to compare against MIN_VALUE as this is u8 and 0 is the MIN_VALUE.
+        *bits <= Self::MAX_VALUE
+    }
+}
+
+// Safety: The enum is `repr(u8)`. All values are `u8` and fall within
+// the min and max values.
+unsafe impl bytemuck::Contiguous for ColorSpaceTag {
+    type Int = u8;
+    const MIN_VALUE: u8 = Self::Srgb as u8;
+    const MAX_VALUE: u8 = Self::XyzD65 as u8;
+}
+
+// Safety: The enum is `repr(u8)` and has only fieldless variants.
+unsafe impl bytemuck::NoUninit for HueDirection {}
+
+// Safety: The enum is `repr(u8)` and `0` is a valid value.
+unsafe impl bytemuck::Zeroable for HueDirection {}
+
+// Safety: The enum is `repr(u8)`.
+unsafe impl bytemuck::checked::CheckedBitPattern for HueDirection {
+    type Bits = u8;
+
+    fn is_valid_bit_pattern(bits: &u8) -> bool {
+        use bytemuck::Contiguous;
+        // Don't need to compare against MIN_VALUE as this is u8 and 0 is the MIN_VALUE.
+        *bits <= Self::MAX_VALUE
+    }
+}
+
+// Safety: The enum is `repr(u8)`. All values are `u8` and fall within
+// the min and max values.
+unsafe impl bytemuck::Contiguous for HueDirection {
+    type Int = u8;
+    const MIN_VALUE: u8 = Self::Shorter as u8;
+    const MAX_VALUE: u8 = Self::Decreasing as u8;
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::{AlphaColor, OpaqueColor, PremulColor, PremulRgba8, Rgba8, Srgb};
-    use bytemuck::{TransparentWrapper, Zeroable};
+    use crate::{
+        AlphaColor, ColorSpaceTag, HueDirection, OpaqueColor, PremulColor, PremulRgba8, Rgba8, Srgb,
+    };
+    use bytemuck::{checked::try_from_bytes, Contiguous, TransparentWrapper, Zeroable};
     use core::marker::PhantomData;
 
     fn assert_is_pod(_pod: impl bytemuck::Pod) {}
@@ -109,6 +164,41 @@ mod tests {
         assert_is_pod(a);
     }
 
+    #[test]
+    fn checked_bit_pattern() {
+        let valid = bytemuck::bytes_of(&2_u8);
+        let invalid = bytemuck::bytes_of(&200_u8);
+
+        assert_eq!(
+            Ok(&ColorSpaceTag::Lab),
+            try_from_bytes::<ColorSpaceTag>(valid)
+        );
+
+        assert!(try_from_bytes::<ColorSpaceTag>(invalid).is_err());
+
+        assert_eq!(
+            Ok(&HueDirection::Increasing),
+            try_from_bytes::<HueDirection>(valid)
+        );
+
+        assert!(try_from_bytes::<HueDirection>(invalid).is_err());
+    }
+
+    #[test]
+    fn contiguous() {
+        let cst1 = ColorSpaceTag::LinearSrgb;
+        let cst2 = ColorSpaceTag::from_integer(cst1.into_integer());
+        assert_eq!(Some(cst1), cst2);
+
+        assert_eq!(None, ColorSpaceTag::from_integer(255));
+
+        let hd1 = HueDirection::Decreasing;
+        let hd2 = HueDirection::from_integer(hd1.into_integer());
+        assert_eq!(Some(hd1), hd2);
+
+        assert_eq!(None, HueDirection::from_integer(255));
+    }
+
     // If the inner type is wrong in the unsafe impl above,
     // that will result in failures here due to assertions
     // within bytemuck.
@@ -148,5 +238,11 @@ mod tests {
                 a: 0
             }
         );
+
+        let cst = ColorSpaceTag::zeroed();
+        assert_eq!(cst, ColorSpaceTag::Srgb);
+
+        let hd = HueDirection::zeroed();
+        assert_eq!(hd, HueDirection::Shorter);
     }
 }